Move needle server api to server generator
