+++ generators/server/internal/needle-api/needle-server-liquibase.js
@@ -18,7 +18,7 @@
  */
 const chalk = require('chalk');
 const needleServer = require('./needle-server');
-const constants = require('../../generator-constants');
+const constants = require('../../../generator-constants');
 
 const SERVER_MAIN_RES_DIR = constants.SERVER_MAIN_RES_DIR;
